Update style.css
The realization that the 'article' tag does not exist within layout builder was a serious blow to mining production. The amount of vespian gas has increased three fold now that the change has been made. We should all be grateful (full of grate?) that the layout builder is as it is.