Hero unit work
Fixed the hero unit videos to work/fit properly in the new layout builder settings.
Added "Size Priority" as an option to hero units so that we can have 100vh sections.
Video now uses the 100vh by default because otherwise it's crazy ugly (can be easily changed now that I've refactored some of the css)