Block Title Fixes
Moving of the block heading style class to the proper places allows the increase based on options picked (hero/supersize) to work correctly and not become gigantic.

Added in the correct normal/bold options if hero strong or supersize bold is chosen.

By default the hero and supersize should not be bolded.