Sidebar spacing fixes
Fixed sidebar spacing so that things vertically aligned.
Also solves the problem of a mobile horizontal scroll bar appearing because of the `g-0` class.
Changed the column sizes to match the same sizes as layout builder sidebars.
Finally fixed alignment of multi-column layoutbuilders + a sidebar. Though this should be a rare case, the new changes make it so all the content aligns as expected.