Interim footer fixes
These changes should not be merged in without Jacob's updates.
Need to double check that the fixes to the footer don't break everything else related to menus.

This will someone revert Tim's changes from PR #945
https://github.com/CuBoulder/tiamat-theme/issues/945

The problem in that ticked was fixed by the changes made but it also removed the column nature of the footer. I had introduced a display error without realizing it by doing `{{ item.content }}` instead of just `{{ item }}`

This should fix everything related to footer spacing and block arrangement.