Update region--site-information.html.twig
Removal of the `ucb-home-link` as the `{{ content }}` generates that link properly for the homepage.

This also means we do not hide the Site Contact Info Footer in the Block Layout. The homepage is currently configured this way so that's no worry there. When they add social media links they'll just need to make sure that it is the first item in the Site Information region to stack properly.

Updated various links to the new ones provided by Wendy